+++ app/android/src/main/java/app/campfire/android/di/ActivityComponent.kt
@@ -6,8 +6,6 @@ package app.campfire.android.di
 import android.app.Activity
 import androidx.core.os.ConfigurationCompat
 import app.campfire.account.api.UserSessionManager
-import app.campfire.audioplayer.cast.CastController
-import app.campfire.audioplayer.impl.MediaControllerConnector
 import app.campfire.audioplayer.offline.OfflineDownloadManager
 import app.campfire.common.root.CampfireContent
 import app.campfire.core.ComponentActivityPlugin
@@ -26,8 +24,6 @@ import me.tatarka.inject.annotations.Provides
 )
 interface ActivityComponent {
   val campfireContent: CampfireContent
-  val mediaControllerConnector: MediaControllerConnector
-  val castController: CastController
   val componentActivityPlugins: Set<ComponentActivityPlugin>
   val offlineDownloadManager: OfflineDownloadManager
   val userSessionManager: UserSessionManager